add component for keyboard avoiding (#5)
