+++ santoku-app/src/app/components/basics/index.ts
@@ -4,6 +4,7 @@ import TextButton from './TextButton';
 import Container from './Container';
 import Text from './Text';
 import Section from './Section';
+import KeyboardAvoidingView from './KeyboardAvoidingView';
 import { translateToViewData } from './ViewUtilities';
 
-export { Description, Title, Text, TextButton, Container, Section, translateToViewData };
+export { Description, Title, Text, TextButton, Container, Section, KeyboardAvoidingView, translateToViewData };
